Fixed an import. Now figure out jest config to work with new index files
