+++ homepairs/homepairsUI/babel.config.js
@@ -34,9 +34,9 @@ module.exports = function(api) {
                         'homepairs-base-styles': './res/Styles/base.ts',
                         'homepairs-components': './src/Components/index.tsx',
                         'homepairs-utilities': './src/utility/index.tsx',
-                        'homepairs-routes' : './src/Routes/RouteConstants',
+                        'homepairs-routes' : './src/Routes/index.tsx',
                         'homepairs-endpoints' : './src/Routes/RemoteEndpoints.tsx',
-                        'homepairs-test' : './tests/fixtures/index.tsx',
+                        'homepairs-test' : './tests/fixtures/index',
                     },
                 },
             ],
