--- homepairs/homepairsUI/src/utility/SyntaxVerification/EmailSyntaxVerification.tsx
@@ -1,24 +1,0 @@
-/* eslint-disable no-useless-escape */
-/* eslint-disable no-control-regex */
-
-const expression = /(?!.*\.{2})^([a-z\d!#$%&'*+\-\/=?^_`{|}~\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF]+(\.[a-z\d!#$%&'*+\-\/=?^_`{|}~\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF]+)*|"((([ \t]*\r\n)?[ \t]+)?([\x01-\x08\x0b\x0c\x0e-\x1f\x7f\x21\x23-\x5b\x5d-\x7e\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF]|\\[\x01-\x09\x0b\x0c\x0d-\x7f\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF]))*(([ \t]*\r\n)?[ \t]+)?")@(([a-z\d\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF]|[a-z\d\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF][a-z\d\-._~\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF]*[a-z\d\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])\.)+([a-z\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF]|[a-z\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF][a-z\d\-._~\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF]*[a-z\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])\.?$/i;
-
-/**
- * -------------------------------------------------
- * Email Syntax Valid 
- * -------------------------------------------------
- * @param {string} email 
- * Returns a boolean dependant of whether a string matches the 
- * valid email format. 
- * 
- * Examples: 
- * aaa@gmail.com 
- * 394;0@gmail.com
- * jimBrown@hotmail.live.com
- * 
- */
-export default function isEmailSyntaxValid(email: string) : Boolean {
-    // TO DO: verify email
-    // how to work with strings in JS https://www.digitalocean.com/community/tutorials/how-to-work-with-strings-in-javascript
-        return expression.test(email);
-}
