+++ homepairs/homepairsUI/tsconfig.jest.json
@@ -18,7 +18,7 @@
         "homepairs-components": "./src/Screens/Components/index.tsx",
         "homepairs-modals": "./src/Modals/index.tsx",
         "homepairs-utilities": "./src/utility/index.tsx",
-        "homepairs-routes" : "./src/Routes/RouteConstants.tsx",
-        "homepairs-test" : "./tests/fixtures/index.tsx",
+        "homepairs-routes" : "./src/Routes/index.tsx",
+        "homepairs-test" : "./tests/fixtures/index",
     },
 }
